requirements.txt
Django==1.7
Werkzeug==0.9.6
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-dajaxice==0.6
django-extensions==1.4.0
django-filter==0.8
django-getenv==1.3.1
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
simplejson==3.6.3
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
